requirements.txt

streamlit==1.37.1
python-dotenv==1.0.1
langchain-openai==0.1.23
openai==1.51.2
pandas==2.2.2
